+ poprawki; +organizacja plików; dodane modele/DTO/serwisy/kontrolery dla części kompendium;

+++ src/backend/TtrpgHelperBackend/TtrpgHelperBackend/Services/Resource/NpcService.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using TtrpgHelperBackend.DTOs.Resource;
+using TtrpgHelperBackend.DTOs.Resource.Npc;
 using TtrpgHelperBackend.Models.Resource;
 
 
@@ -7,15 +7,27 @@ namespace TtrpgHelperBackend.Services.Resource;
 
 public interface INpcService
 {
+    // ==============
+    // -- CAMPAIGN --
     Task<GetScenarioNpcDto?> GetNpc(int npcId, int gameMasterId);
     Task<IEnumerable<GetScenarioNpcDto>> GetNpcs(int campaignId, int gameMasterId);
     Task<GetScenarioNpcDto?> CreateNpc(CreateNpcDto dto, int gameMasterId);
     Task<GetScenarioNpcDto?> UpdateNpc(UpdateNpcDto dto, int gameMasterId);
     Task<bool> DeleteNpc(int npcId, int gameMasterId);
     
+    // ==========
+    // -- BOTH --
     Task<GetScenarioNpcSkillDto?> AddNpcSkill(int npcId, CreateNpcSkillDto dto, int gameMasterId);
     Task<GetScenarioNpcSkillDto?> UpdateNpcSkill(UpdateNpcSkillDto dto, int gameMasterId);
     Task<bool> DeleteNpcSkill(int npcSkillId, int gameMasterId);
+    
+    // ================
+    // -- COMPENDIUM --
+    Task<GetScenarioNpcDto?> GetCompendiumNpc(int npcId);
+    Task<IEnumerable<GetScenarioNpcDto>> GetCompendiumNpcs();
+    Task<GetScenarioNpcDto?> CreateCompendiumNpc(CreateCompendiumNpcDto dto);
+    Task<GetScenarioNpcDto?> UpdateCompendiumNpc(UpdateNpcDto dto);
+    Task<bool> DeleteCompendiumNpc(int npcId);
 }
 
 public class NpcService : INpcService
@@ -27,11 +39,13 @@ public class NpcService : INpcService
         _db = db;
     }
     
+    // ==============
+    // -- CAMPAIGN --
     public async Task<GetScenarioNpcDto?> GetNpc(int npcId, int gameMasterId)
     {
         var npc = await IncludeAllNpcData()
             .FirstOrDefaultAsync(n => n.Id == npcId);
-        if (npc == null || npc.Campaign.GameMasterId != gameMasterId) return null;
+        if (npc == null || npc.Campaign == null || npc.Campaign.GameMasterId != gameMasterId) return null;
 
         return NpcDto(npc);
     }
@@ -39,7 +53,7 @@ public class NpcService : INpcService
     public async Task<IEnumerable<GetScenarioNpcDto>> GetNpcs(int campaignId, int gameMasterId)
     {
         var npcs = await IncludeAllNpcData()
-            .Where(n => n.CampaignId == campaignId && n.Campaign.GameMasterId == gameMasterId)
+            .Where(n => n.CampaignId == campaignId && n.Campaign != null &&  n.Campaign.GameMasterId == gameMasterId)
             .ToListAsync();
 
         return npcs.Select(NpcDto).ToList();
@@ -77,7 +91,7 @@ public class NpcService : INpcService
     {
         var npc = await IncludeAllNpcData()
             .FirstOrDefaultAsync(n => n.Id == dto.Id);
-        if (npc == null || npc.Campaign.GameMasterId != gameMasterId) return null;
+        if (npc == null || npc.Campaign == null || npc.Campaign.GameMasterId != gameMasterId) return null;
 
         npc.Name = dto.Name;
         npc.Description = dto.Description;
@@ -101,7 +115,7 @@ public class NpcService : INpcService
         var npc = await _db.Npcs
             .Include(n => n.Campaign)
             .FirstOrDefaultAsync(n => n.Id == npcId);
-        if (npc == null || npc.Campaign.GameMasterId != gameMasterId) return false;
+        if (npc == null || npc.Campaign == null || npc.Campaign.GameMasterId != gameMasterId) return false;
 
         _db.Npcs.Remove(npc);
         await _db.SaveChangesAsync();
@@ -109,12 +123,15 @@ public class NpcService : INpcService
         return true;
     }
     
+    
+    // ==========
+    // -- BOTH --
     public async Task<GetScenarioNpcSkillDto?> AddNpcSkill(int npcId, CreateNpcSkillDto dto, int gameMasterId)
     {
         var npc = await _db.Npcs
             .Include(n => n.Campaign)
             .FirstOrDefaultAsync(n => n.Id == npcId);
-        if (npc == null || npc.Campaign.GameMasterId != gameMasterId) return null;
+        if (npc == null || npc.Campaign == null || npc.Campaign.GameMasterId != gameMasterId) return null;
 
         var skill = new NpcSkill
         {
@@ -137,7 +154,7 @@ public class NpcService : INpcService
                 .ThenInclude(n => n.Campaign)
             .FirstOrDefaultAsync(s => s.Id == dto.Id);
 
-        if (skill == null || skill.Npc.Campaign.GameMasterId != gameMasterId) return null;
+        if (skill == null || skill.Npc.Campaign == null || skill.Npc.Campaign.GameMasterId != gameMasterId) return null;
 
         skill.Name = dto.Name;
         skill.Description = dto.Description;
@@ -154,7 +171,7 @@ public class NpcService : INpcService
             .Include(s => s.Npc)
                 .ThenInclude(n => n.Campaign)
             .FirstOrDefaultAsync(s => s.Id == npcSkillId);
-        if (skill == null || skill.Npc.Campaign.GameMasterId != gameMasterId) return false;
+        if (skill == null || skill.Npc.Campaign == null || skill.Npc.Campaign.GameMasterId != gameMasterId) return false;
 
         _db.NpcSkills.Remove(skill);
         await _db.SaveChangesAsync();
@@ -162,6 +179,87 @@ public class NpcService : INpcService
         return true;
     }
 
+    
+    // ================
+    // -- COMPENDIUM --
+    public async Task<GetScenarioNpcDto?> GetCompendiumNpc(int npcId)
+    {
+        var npc = await _db.Npcs
+            .FirstOrDefaultAsync(n => n.Id == npcId && n.IsCompendium);
+        if (npc == null) return null;
+
+        return NpcDto(npc);
+    }
+
+    public async Task<IEnumerable<GetScenarioNpcDto>> GetCompendiumNpcs()
+    {
+        var npcs = await _db.Npcs
+            .Where(n => n.IsCompendium)
+            .ToListAsync();
+        
+        return npcs.Select(NpcDto).ToList();
+    }
+
+    public async Task<GetScenarioNpcDto?> CreateCompendiumNpc(CreateCompendiumNpcDto dto)
+    {
+        var npc = new Npc
+        {
+            CampaignId = null,
+            Name = dto.Name,
+            Description = dto.Description,
+            RaceId = dto.RaceId,
+            ClassId = dto.ClassId,
+            Level = dto.Level,
+            Strength = dto.Strength,
+            Dexterity = dto.Dexterity,
+            Constitution = dto.Constitution,
+            Intelligence = dto.Intelligence,
+            Wisdom = dto.Wisdom,
+            Charisma = dto.Charisma
+        };
+        
+        _db.Npcs.Add(npc);
+        await _db.SaveChangesAsync();
+        
+        return NpcDto(npc);
+    }
+
+    public async Task<GetScenarioNpcDto?> UpdateCompendiumNpc(UpdateNpcDto dto)
+    {
+        var npc = await _db.Npcs
+            .FirstOrDefaultAsync(n => n.Id == dto.Id &&  n.IsCompendium);
+        if (npc == null) return null;
+        
+        npc.Name = dto.Name;
+        npc.Description = dto.Description;
+        npc.RaceId = dto.RaceId;
+        npc.ClassId = dto.ClassId;
+        npc.Level = dto.Level;
+        npc.Strength = dto.Strength;
+        npc.Dexterity = dto.Dexterity;
+        npc.Constitution = dto.Constitution;
+        npc.Intelligence = dto.Intelligence;
+        npc.Wisdom = dto.Wisdom;
+        npc.Charisma = dto.Charisma;
+        
+        await _db.SaveChangesAsync();
+        
+        return NpcDto(npc);
+    }
+
+    public async Task<bool> DeleteCompendiumNpc(int npcId)
+    {
+        var npc = await _db.Npcs
+            .FirstOrDefaultAsync(n => n.Id == npcId &&  n.IsCompendium);
+        if (npc == null) return false;
+        
+        _db.Npcs.Remove(npc);
+        await _db.SaveChangesAsync();
+        
+        return true;
+    }
+    
+    
     // ====================
     // -- HELPER METHODS --
     private IQueryable<Npc> IncludeAllNpcData()
